% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt_spas_x_bayes_params}
\alias{bt_spas_x_bayes_params}
\title{BT-SPAS-X BUGS parameters}
\format{
A named list with 4 elements:
\itemize{
\item \code{number_mcmc}: TODO
\item \code{number_burnin}:
\item \code{number_thin}:
\item \code{number_chains}:
}
}
\usage{
bt_spas_x_bayes_params
}
\description{
Parameters for running BUGS on the BT-SPAS-X model
}
\keyword{datasets}
