% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{bt_spas_x_bugs}
\alias{bt_spas_x_bugs}
\title{Execute BT-SPAS-X in WinBUGs}
\usage{
bt_spas_x_bugs(
  data,
  inits,
  parameters,
  model_name,
  bt_spas_x_bayes_params,
  number_mcmc,
  bugs_directory
)
}
\arguments{
\item{data}{a list containing the following elements:
\itemize{
\item \strong{Nmr} number of unique mark-recapture experiments performed across tributaries, years and weeks
\item \strong{Ntribs} number of tributaries to use for the pCap component of the model
\item \strong{Nstrata} number of weeks with catch data
\item \strong{Uwc_ind} number of weeks in catch data with associated pCap and flow data
\item \strong{Nwomr} number of weeks in catch data without associated pCap and flow data
\item \strong{Nstrata_wc} number of weeks in catch data with catch data (RST fished)
\item \strong{indices_site_pCap} tributary index (of all possible tributaries) to use to predict efficiency for missing strata. Note length=0 if selected tributary is not part of trib set that has mark-recap data. In this case model that samples from trib hyper will be called.
\item \strong{ind_trib} indices (1:Ntribs) assigned to the mark-recapture experiment table for use in BUGs
\item \strong{ind_pCap} indices of weeks in mark-recapture table for U strata being estimated
\item \strong{Uind_woMR} indices of weeks in catch data without associated pCap and flow data
\item \strong{Uind_wMR} indices of weeks in catch data with associated pCap and flow data
\item \strong{Uwc_ind} indices of weeks in catch data with catch data
\item \strong{Releases} number of fish released for each mark-recapture experiment
\item \strong{Recaptures} number of fish recaptured for each mark-recapture experiment
\item \strong{u} weekly abundance
\item \strong{mr_flow} standardized flow, averaged over recapture days (< 1 week)
\item \strong{catch_flow} standardized flow, averaged by week
\item \strong{K} number of columns in the bspline basis matrix
\item \strong{ZP} The b_spline_matrix (bspline basis matrix. One row for each data point (1:Nstrata), and one column for each term in the cubic polynomial function (4) + number of knots)
\item \strong{lgN_max} maxmimum possible value for log N across strata
}}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}.}
}
\value{
either a list of the required inputs for a WinBUGS model (if running on a Mac), or the results
of the model run.
}
\description{
Execute BT-SPAS-X in WinBUGs
}
\details{
This function is called within \code{run_single_bt_spas_x()} and calls the WinBUGS code
}
