% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{calculate_ss_tot}
\alias{calculate_ss_tot}
\title{Passage to Spawner Sum of Squares}
\usage{
calculate_ss_tot(data)
}
\description{
This function calculates the total sum of squares, which is required
as input to the \code{data} call for \code{fit_passage_to_spawner_model()}.
}
\keyword{internal}
