% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{compare_P2S_covariates_within_stream}
\alias{compare_P2S_covariates_within_stream}
\title{Compare Environmental Covariates in Passage to Spawner (P2S) Model Within a Stream}
\usage{
compare_P2S_covariates_within_stream(data, stream_name)
}
\value{
A table containing the following variables:
\itemize{
\item \strong{par_names} Parameter name
\item \strong{stream} Mean of the posterior distribution for a parameter
\item \strong{mean} Monte Carlo standard error for summary of all chains merged. Seee \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details}
\item \strong{median} 50\% quantile of posterior distribution for a parameter.
\item \strong{sd} standard deviation
\item \strong{covar_considered} the covariate considered
}
}
\description{
This function is called within \code{compare_P2S_model_covariates()} and iterates through all
environmental covariates for a stream, producing diagnostic statistics for that stream.
}
\seealso{
Other passage_to_spawner: 
\code{\link{compare_P2S_model_covariates}()},
\code{\link{run_passage_to_spawner_model}()}
}
\concept{passage_to_spawner}
\keyword{internal}
