% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{compare_P2S_model_covariates}
\alias{compare_P2S_model_covariates}
\title{Compare Environmental Covariates in Passage to Spawner (P2S) Model}
\usage{
compare_P2S_model_covariates(data, stream_name)
}
\value{
A table containing the following variables:
\itemize{
\item \strong{par_names} Parameter name
\item \strong{stream} Mean of the posterior distribution for a parameter
\item \strong{mean} Monte Carlo standard error for summary of all chains merged (see \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details})
\item \strong{median} 50\% quantile of posterior distribution for a parameter.
\item \strong{sd} standard deviation
\item \strong{covar_considered} the covariate considered
}
}
\description{
This function is called in \code{run_passage_to_spawner_model()} and iterates
through available environmental covariates for a stream, pulling out diagnostic statistics. The
data is truncated to only include years for which all environmental covariates are available.
}
\seealso{
Other passage_to_spawner: 
\code{\link{run_passage_to_spawner_model}()}
}
\concept{passage_to_spawner}
