% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{diagnostic_plots_split}
\alias{diagnostic_plots_split}
\title{BT SPAS X diagnostic plots}
\usage{
diagnostic_plots_split(site_arg, run_year_arg, abundance_model)
}
\arguments{
\item{site_arg}{The site being fit}

\item{run_year_arg}{The run year being fit}

\item{abundance_model}{The STANfit model object produced by running \code{fit_abundance_model()}}
}
\value{
A plot
}
\description{
BT SPAS X diagnostic plots
}
\details{
This function produces a plot with data and results of fitting the pCap and abundance models.
}
