% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{.version_from_blob_url}
\alias{.version_from_blob_url}
\title{Parses the pins version token out of a stored blob URL:
.../model-fits/<model_name>/<model_name>/\if{html}{\out{<version>}}/<model_name>.rds}
\usage{
.version_from_blob_url(url)
}
\description{
Parses the pins version token out of a stored blob URL:
.../model-fits/<model_name>/<model_name>/\if{html}{\out{<version>}}/<model_name>.rds
}
\keyword{internal}
