% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{extract_P2S_estimates}
\alias{extract_P2S_estimates}
\title{Extract parameter estimates from P2S}
\usage{
extract_P2S_estimates(passage_to_spawner_model_object)
}
\value{
a table with the following variables:
\itemize{
\item \strong{parameter} Parameter name
\item \strong{mean} Mean of the posterior distribution for a parameter
\item \strong{se_mean} Monte Carlo standard error for summary of all chains merged (see \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details})
\item \strong{sd} Standard deviation of the posterior distribution for a parameter
\item \strong{\code{2.5}} 2.5\% quantile of posterior distribution for a parameter.
\item \strong{\code{25}} 25\% quantile of posterior distribution for a parameter.
\item \strong{\code{50}} 50\% quantile of posterior distribution for a parameter.
\item \strong{\verb{7\%}} 75\% quantile of posterior distribution for a parameter.
\item \strong{\code{97.5}} 97.5\% quantile of posterior distribution for a parameter.
\item \strong{n_eff} Effective sample size for a parameter
\item \strong{Rhat} Split Rhats for a parameter
\item \strong{year} Year observed data came from
}
}
\description{
This function takes in a STAN fit object produced by running \code{run_passage_to_spawner_model()} and produces a formatted
table of parameter estimates from the Passage to Spawner model.
}
\seealso{
Other passage_to_spawner: 
\code{\link{compare_P2S_model_covariates}()},
\code{\link{run_passage_to_spawner_model}()}
}
\concept{passage_to_spawner}
