% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{extract_stock_recruit_estimates}
\alias{extract_stock_recruit_estimates}
\title{Extract stock-recruit parameter estimates}
\usage{
extract_stock_recruit_estimates(
  stock_recruit_inputs,
  stock_recruit_model_object
)
}
\arguments{
\item{stock_recruit_inputs}{a named list produced by running \code{prepare_stock_recruit_inputs()}}

\item{stock_recruit_model_object}{a STANfit object produced by running \code{fit_stock_recruit_model()}}
}
\value{
a table with the following variables:
\itemize{
\item \strong{parameter} Parameter name
\item \strong{mean} Mean of the posterior distribution for a parameter
\item \strong{se_mean} Monte Carlo standard error for summary of all chains merged (see \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details})
\item \strong{sd} Standard deviation of the posterior distribution for a parameter
\item \strong{\code{2.5}} 2.5\% quantile of posterior distribution for a parameter.
\item \strong{\code{25}} 25\% quantile of posterior distribution for a parameter.
\item \strong{\code{50}} 50\% quantile of posterior distribution for a parameter.
\item \strong{\verb{7\%}} 75\% quantile of posterior distribution for a parameter.
\item \strong{\code{97.5}} 97.5\% quantile of posterior distribution for a parameter.
\item \strong{n_eff} Effective sample size for a parameter
\item \strong{Rhat} Split Rhats for a parameter
\item \strong{year} Year observed data came from
}
}
\description{
This function takes in a STAN fit object produced by running \code{fit_stock_recruit_model()} and produces a formatted
table of parameter estimates.
}
\details{
This function extracts parameter estimates from the stock recruit model object.
}
