% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{extract_survival_estimates}
\alias{extract_survival_estimates}
\title{Extract parameter estimates from the survival STAN model object}
\usage{
extract_survival_estimates(model_object)
}
\arguments{
\item{model_object}{the STAN object produced by running \code{fit_survival_model()}.}
}
\value{
A table with the format:
\itemize{
\item \strong{parameter}
\item \strong{year}
\item \strong{reach_name}
\item \strong{release_group}
\item \strong{statistic}
\item \strong{value}
\item \strong{srjpedata_version}
}
}
\description{
Extract parameter estimates from the survival STAN model object
}
\details{
This function extracts parameter estimates from the survival STAN model and returns them in a tidy data frame format.
}
