% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{fit_abundance_model}
\alias{fit_abundance_model}
\title{Fit abundance model}
\usage{
fit_abundance_model(input, pCap_fit)
}
\arguments{
\item{input}{A list containing the inputs for the abundance model.}

\item{pCap_fit}{A STANfit object resulting from running \code{fit_pCap_model()}}
}
\value{
a named list with the required elements for that model run.
}
\description{
Fit abundance model
}
\details{
This function prepares the data list for the abundance STAN model based on what the model name is.
}
\keyword{internal}
