% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{fit_stock_recruit_model}
\alias{fit_stock_recruit_model}
\title{Fit stock-recruit model in STAN}
\usage{
fit_stock_recruit_model(stock_recruit_inputs)
}
\arguments{
\item{stock_recruit_inputs}{the object produced by \code{prepare_stock_recruit_inputs()}}
}
\value{
a STAN object.
}
\description{
Fit stock-recruit model in STAN
}
\details{
This function runs the STAN stock-recruit model.
}
\seealso{
Other Fit model:
\code{\link[=fit_abundance_model_BUGS]{fit_abundance_model_BUGS()}},
\code{\link[=fit_inseason_model]{fit_inseason_model()}},
\code{\link[=fit_pCap_model]{fit_pCap_model()}},
\code{\link[=fit_passage_to_spawner_model]{fit_passage_to_spawner_model()}},
\code{\link[=fit_survival_model]{fit_survival_model()}}
}
\concept{Fit model}
