% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{generate_diagnostic_plot}
\alias{generate_diagnostic_plot}
\title{Diagnostic plots}
\usage{
generate_diagnostic_plot(inputs, fit)
}
\arguments{
\item{inputs}{inputs for the fit model, created by running the appropriate \code{prepare_inputs} function.}

\item{fit}{the model fit}

\item{model_results_name}{The model type name associated with options in the database: either
\code{all_mark_recap}, \code{no_mark_recap}, \code{missing_mark_recap}, \code{no_mark_recap_no_trib},
\code{pcap_all}, \code{pcap_mainstem}, \code{p2s}, or \code{stock_recruit}.}
}
\value{
Generates a diagnostic posterior predictive plot for a model fit.
}
\description{
Diagnostic plots
}
\details{
This function produces a posterior predictive check plot.
}
