% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{generate_diagnostic_plot_juv}
\alias{generate_diagnostic_plot_juv}
\title{BT SPAS X diagnostic plots}
\usage{
generate_diagnostic_plot_juv(inputs, con)
}
\arguments{
\item{inputs}{a list of inputs generated by running \code{prepare_abundance_inputs()}}

\item{con}{a valid connection to the database.}
}
\value{
A plot.
}
\description{
BT SPAS X diagnostic plots
}
\details{
This function produces a plot with data and results of fitting the pCap and abundance models for
a given site and run year.
}
