% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{generate_diagnostic_plot_p2s}
\alias{generate_diagnostic_plot_p2s}
\title{P2S diagnostic plots}
\usage{
generate_diagnostic_plot_p2s(inputs, fit, local_folder)
}
\arguments{
\item{inputs}{inputs for the fit model, created by running \code{prepare_P2S_inputs()}}

\item{fit}{the model fit}

\item{local_folder}{the filepath to the local folder, ending in \code{/}, where you want to store the figures.}
}
\value{
Saves two plots to your local filepath.
}
\description{
P2S diagnostic plots
}
\details{
This function prodces two basic diagnostic plots and saves them
to a local filepath.
}
