% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{generate_diagnostic_plot_sr_covar}
\alias{generate_diagnostic_plot_sr_covar}
\title{Stock-recruit diagnostic plots}
\usage{
generate_diagnostic_plot_sr_covar(sr_inputs, sr_fit)
}
\arguments{
\item{sr_inputs}{inputs for the fit model, created by running \code{prepare_stock_recruit_inputs()}}

\item{sr_fit}{the model fit object, created by running \code{fit_stock_recruit_model()}}
}
\value{
A plot.
}
\description{
Stock-recruit diagnostic plots
}
\details{
This function produces a plot with observed spawner and estimated recruitment data,
alongside predicted recruitment across a range of covariate values produced using
posterior estimates from the model fit. It is meant to compare predicted and observed
recruitment as a function of spawner effect on covariate.
}
