% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_season_model.R
\name{generate_results_plot_inseason}
\alias{generate_results_plot_inseason}
\title{Inseason forecast plot}
\usage{
generate_results_plot_inseason(inseason_inputs, con)
}
\arguments{
\item{con}{A database connection object.}

\item{forecast_inputs}{A list with forecast metadata and inputs, including For_ewk and calendar labels.}
}
\value{
A ggplot object.
}
\description{
This function produces a plot of observed and predicted abundance
using posterior summaries retrieved from the database.
}
