% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{generate_results_plot_p2s}
\alias{generate_results_plot_p2s}
\title{P2S results plots}
\usage{
generate_results_plot_p2s(p2s_inputs, con)
}
\arguments{
\item{p2s_inputs}{inputs for the fit model, created by running \code{prepare_P2S_inputs()}}

\item{con}{a connection to the database.}
}
\value{
A plot showing observed and predicted spawners by year.
}
\description{
P2S results plots
}
\details{
This function produces a plot with predicted spawners with uncertainty.
for simple comparison.
}
