% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{generate_survival_rate_plot}
\alias{generate_survival_rate_plot}
\title{Plot smolt survival rate}
\usage{
generate_survival_rate_plot(survival_estimates)
}
\arguments{
\item{survival_estimates}{A table that is produced from running \code{extract_survival_estimates()}.}
}
\value{
A plot.
}
\description{
Plot smolt survival rate
}
\details{
This function produces a plot showing the parameter estimates for \code{SurvRelSac} and \code{SurvWoodSac}, which
are survival rates for Release to Woodson Bridge and Woodson Bridge to Sacramento.
}
