% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{get_abundance_data_list}
\alias{get_abundance_data_list}
\title{Prepare Data Object for abundance STAN call}
\usage{
get_abundance_data_list(full_data_list)
}
\arguments{
\item{full_data_list}{a list containing all possible data objects to use in the model.}
}
\value{
a named list with the required elements for that model run.
}
\description{
Prepare Data Object for abundance STAN call
}
\details{
This function prepares the data list for the abundance STAN model based on what the model name is.
}
\keyword{internal}
