% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{get_all_pars}
\alias{get_all_pars}
\title{Extract Passage to Spawner Estimates}
\usage{
get_all_pars(model_fit, stream_name)
}
\description{
This function extracts parameter estimates and summary statistics for
the Passage to Spawner model (see \code{?run_passage_to_spawner_model}). The function
calls \code{summary()} on the \code{stanfit} object produced by running the model. For details,
see \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details}.
}
\keyword{internal}
