% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_hierarchical_parameter_estimates}
\alias{get_hierarchical_parameter_estimates}
\title{Extract hierarchical parameters from BT-SPAS-X}
\usage{
get_hierarchical_parameter_estimates(model_fit_object)
}
\description{
TODO
}
\keyword{internal}
