% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_model_object}
\alias{get_model_object}
\title{Get Model Object}
\usage{
get_model_object(
  con,
  keyword = NULL,
  model_run_id = NULL,
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY")
)
}
\arguments{
\item{con}{A connection object to the database.}

\item{keyword}{An optional string used to search for the model run description in the database. If provided, it is used to find the corresponding blob URL.}

\item{model_run_id}{An optional ID for a specific model run. If provided, it is used to find the corresponding blob URL. If neither \code{keyword} nor \code{model_run_id} is provided, the latest model run is used.}

\item{access_key}{A string specifying the Azure storage access key with read permissions. By default, it retrieves from the environment variable \code{AZ_CONTAINER_ACCESS_KEY}.}
}
\value{
The model object retrieved from the Azure Blob Storage.
#' @examples
\dontrun{
model_object <- get_model_object(con = db_connection, keyword = "model description")

# Example: Get a model object from Azure Blob Storage using a model run ID
model_object <- get_model_object(con = db_connection, model_run_id = 12)

print(model_object$model.file)
}
}
\description{
This function retrieves the model object (.Rds file) from Azure Blob Storage using a keyword from the model run description or a model run ID.
}
