% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_model_results_parameters}
\alias{get_model_results_parameters}
\title{Get Model Parameter Results}
\usage{
get_model_results_parameters(con, keyword = NULL, model_run_id = NULL)
}
\arguments{
\item{con}{A connection object to the database.}

\item{keyword}{An optional string used to search for the model run description in the database. If provided, it is used to find the corresponding blob URL.}

\item{model_run_id}{An optional ID for a specific model run. If provided, it is used to find the corresponding blob URL. If neither \code{keyword} nor \code{model_run_id} is provided, the latest model run is used.}
}
\value{
A tibble containing model parameters.
}
\description{
This function retrieves the model parameter results from the JPE Database using a keyword(s) from the model run description or a model run ID.
}
\examples{
\dontrun{
con <- dbConnect(RPostgres::Postgres(),
                 dbname = "your_db_name",
                 host = "your_host",
                 port = 5432,
                 user = "your_username",
                 password = "your_password")

# Example: get model parameters from Azure JPE Data Storage using a keyword
model_results <- get_model_results_parameters(con, keyword = "model description")
print(model_results)

Example: get model parameters from Azure JPE Data Storage using a model run ID
model_results <- get_model_results_parameters(con, model_run_id = 12)

dbDisconnect(con)
}
}
