% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_most_recent_model_results}
\alias{get_most_recent_model_results}
\title{Get Most Recent Model Results}
\usage{
get_most_recent_model_results(con)
}
\arguments{
\item{con}{A connection object to the database.}
}
\value{
A tibble containing model parameters.
}
\description{
This function retrieves the most recent model results for each model name, site, year, stream, etc.
}
\examples{
\dontrun{
con <- dbConnect(RPostgres::Postgres(),
                 dbname = "your_db_name",
                 host = "your_host",
                 port = 5432,
                 user = "your_username",
                 password = "your_password")

Example: get model parameters from Azure JPE Data Storage
model_results <- get_model_results_parameters(con)

dbDisconnect(con)
}
}
