% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{get_pCap_data_list}
\alias{get_pCap_data_list}
\title{Prepare Data Object for pCap STAN call}
\usage{
get_pCap_data_list(model_name, full_data_list)
}
\arguments{
\item{model_name}{which model to call on the data. Either \code{all_mark_recap.bug},
\code{missing_mark_recap.bug}, \code{no_mark_recap.bug}, or \code{no_mark_recap_no_trib.bug}}

\item{full_data_list}{a list containing all possible data objects to use in the model.}
}
\value{
a named list with the required elements for that model run.
}
\description{
Prepare Data Object for pCap STAN call
}
\details{
This function prepares the data list for a pCap STAN model based on what the model name is.
}
\keyword{internal}
