% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{get_predicted_spawners_from_P2S}
\alias{get_predicted_spawners_from_P2S}
\title{Extract Predicted Spawners from P2S Model}
\usage{
get_predicted_spawners_from_P2S(results_tibble_with_years)
}
\description{
This function is called within \code{run_passage_to_spawner_model()} and pulls only the predicted spawner
counts, 2.5\% and 97.5\% confidence intervals from the \code{stanfit} object.
}
\keyword{internal}
