% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_summary_table_stan}
\alias{get_summary_table_stan}
\title{Extract summary table from stanfit object.}
\usage{
get_summary_table_stan(
  stanfit_object,
  site,
  run_year,
  lifestage,
  weeks_fit,
  sites_fit,
  model_name
)
}
\description{
This function takes in a stanfit object for BT-SPAS-X and extracts parameter
estimates in a clean tibble.
}
