% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_model.R
\name{get_survival_forecast}
\alias{get_survival_forecast}
\title{Get survival forecasts weighted by FL and run size}
\usage{
get_survival_forecast(con)
}
\arguments{
\item{con}{A valid DBI database connection to the SRJPE model run database.}
}
\value{
A tibble containing:
\describe{
\item{\code{site}}{The site for which the survival rate is produced.}
\item{\code{mean_weighted_survival}}{Mean of forecasted survival rate in logit space for outmigrant run for JPE.stan model.}
\item{\code{sd_weighted_survival}}{Standard deviation of forecasted survival rate in logit space for outmigrant run for JPE.stan model.}
}
}
\description{
Pulls together estimates from the inseason, PLAD, and survival models to generate a
forecasted survival rate from RST to Delta that is weighted by fork length and run size.
}
\details{
TODO
}
\concept{Forecast}
