% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{get_years_from_P2S_model_fits}
\alias{get_years_from_P2S_model_fits}
\title{Extract Years Passage to Spawner Estimates}
\usage{
get_years_from_P2S_model_fits(P2S_model_fits)
}
\description{
This function extracts the years the data were collected from the results of
the Passage to Spawner model (see \code{?run_passage_to_spawner_model}).
}
\keyword{internal}
