% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2S_comparison_results}
\alias{P2S_comparison_results}
\alias{P2S_model_fits}
\title{Passage to Spawner Covariate Comparison results}
\format{
A tibble with 140 rows and 9 columns
\itemize{
\item \code{par_names}: Parameter name
\item \code{stream}: Stream for which the model was run
\item \code{mean}: Mean of the posterior distribution for a parameter
\item \code{median}: Median the posterior distribution for a parameter
\item \code{sd`}: Standard deviation of the posterior distribution for a parameter
\item \code{lcl}: 2.5\% lower confidence limit
\item \code{ucl}: 97.5 upper confidence limit
\item \code{covar_considered}: The covariate associated with the results
\item \code{convergence_metric}: Whether or not the convergence metric was reached
}
}
\usage{
P2S_model_fits
}
\description{
Output from running \code{SRJPEmodel::compare_P2S_model_covariates()}
}
\keyword{datasets}
