% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p2s_model_code}
\alias{p2s_model_code}
\title{Passage to Spawner model code}
\format{
An object of class \code{character} of length 1.
}
\usage{
p2s_model_code
}
\description{
A character string containing the STAN model code for the Passage to Spawner submodel.
}
\keyword{datasets}
