% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2S_model_fits}
\alias{P2S_model_fits}
\title{Passage to Spawner Parameter Estimates}
\format{
A tibble with 343 rows and 12 columns
\itemize{
\item \code{par_names}: Parameter name
\item \code{mean}: Mean of the posterior distribution for a parameter
\item \code{se_mean}: Monte Carlo standard error for summary of all chains merged. See \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details}
\item \code{sd}: Standard deviation of the posterior distribution for a parameter
\item \code{2.5\%}: 2.5\% quantile of posterior distribution for a parameter
\item \code{25\%}: 25\% quantile of posterior distribution for a parameter
\item \code{50\%}: 50\% quantile of posterior distribution for a parameter
\item \code{75\%}: 75\% quantile of posterior distribution for a parameter
\item \code{97.5\%}: 97.5\% quantile of posterior distribution for a parameter
\item \code{n_eff}: Effective sample size for a parameter
\item \code{Rhat}: Split Rhats for a parameter
\item \code{stream}: Stream name for the parameter estimate
}
}
\usage{
P2S_model_fits
}
\description{
The passage to spawner model, documented in \code{vignette("passage_to_spawner_submodel", package = "SRJPEmodel")},
estimates the relationship between upstream passage counts and spawner counts (either \code{redd} or \code{holding}) for
several streams. When fit, the model produces an object \code{P2S_model_fits} that contains the parameter names,
mean parameter estimate, confidence intervals, and diagnostics. Currently, this model is only fit for
\verb{battle creek}, \verb{clear creek}, \verb{deer creek}, and \verb{mill creek}.
}
\keyword{datasets}
