% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{plot_juv_data}
\alias{plot_juv_data}
\title{BT SPAS X raw data plots}
\usage{
plot_juv_data(site, run_year)
}
\arguments{
\item{site_arg}{The site being fit}

\item{run_year_arg}{The run year being fit}
}
\value{
A plot
}
\description{
BT SPAS X raw data plots
}
\details{
This function produces a plot with data used to fit BT SPAS X.
}
