% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_juvenile_abundance}
\alias{plot_juvenile_abundance}
\title{Plot weekly estimated juvenile abundance}
\usage{
plot_juvenile_abundance(bt_spas_x_results_clean)
}
\arguments{
\item{bt_spas_x_results_clean:}{the \code{final_results} element from the object produced by running
\code{run_single_bt_spas_x()}}
}
\value{
a plot showing weekly estimated abundance from the model (with confidence intervals)
compared to lincoln-peterson estimates
}
\description{
Plot weekly estimated juvenile abundance
}
\details{
This will generate a bar plot for output of a single run of the bt spas x model.
}
