% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_weekly_capture_probability}
\alias{plot_weekly_capture_probability}
\title{Plot weekly estimated capture probability}
\usage{
plot_weekly_capture_probability(
  bt_spas_x_results_clean,
  mainstem_version = c(F, T)
)
}
\arguments{
\item{bt_spas_x_results_clean:}{the \code{final_results} element from the object produced by running
\code{run_single_bt_spas_x()}}
}
\value{
a plot showing estimated capture probability
}
\description{
Plot weekly estimated capture probability
}
\details{
This will generate a bar plot for output of a single run of the bt spas x model.
}
