% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance_split.R
\name{prepare_inputs_pCap_abundance_STAN}
\alias{prepare_inputs_pCap_abundance_STAN}
\title{Call BT-SPAS-X (WinBUGS) on a single site/run year combination}
\usage{
prepare_inputs_pCap_abundance_STAN(
  weekly_juvenile_abundance_catch_data,
  weekly_juvenile_abundance_efficiency_data,
  site,
  run_year,
  effort_adjust = c(T, F),
  default_lgN_prior_denominator = NULL
)
}
\arguments{
\item{weekly_juvenile_abundance_catch_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_catch_data}}

\item{weekly_juvenile_abundance_efficiency_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_efficiency_data}}

\item{site}{site for which you want to fit the model}

\item{run_year}{run year for which you want to fit the model}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort.}
}
\value{
a list:
\itemize{
\item \strong{pCap_inputs} a list of data and inits for input into the pCap model.
\item \strong{abundance_inputs} a list of data and inits for input into the abundance model.
}
}
\description{
Call BT-SPAS-X (WinBUGS) on a single site/run year combination
}
\details{
This function prepares data for input into a pCap STAN model.
}
