% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{prepare_mainstem_pCap_data}
\alias{prepare_mainstem_pCap_data}
\title{Prepare mainstem inputs for abundance model}
\usage{
prepare_mainstem_pCap_data(mainstem_site)
}
\arguments{
\item{mainstem_site}{site on the mainstem Sacramento River (either \verb{knights landing},
\code{tisdale}, or \verb{red bluff diversion dam}) for which you want to prepare inputs}
}
\value{
a named list, with each element containing the inputs required for one of the three
mainstem sites. In this named list are
\itemize{
\item \strong{inputs} a list of data and inits for input into the abundance model.
\item \strong{sites_fit} The site the inputs were prepared for (either \verb{knights landing},
\code{tisdale}, or \verb{red bluff diversion dam}).
}
}
\description{
Prepare mainstem inputs for abundance model
}
\details{
This function is called within \code{prepare_pCap_inputs()}
}
