% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{prepare_P2S_inputs}
\alias{prepare_P2S_inputs}
\title{Prepare data for P2S}
\usage{
prepare_P2S_inputs(stream, selected_covariate, truncate_dataset = FALSE)
}
\arguments{
\item{truncate_dataset}{a TRUE/FALSE value. If TRUE, will truncate the dataset so that only years where all covariate values are
present are used.}

\item{stream_name:}{The name of the stream for which you would like to run the P2S model. Can be
\verb{battle creek}, \verb{clear creek}, \verb{deer creek}, \verb{mill creek}, or (draft form) \verb{butte creek}.}

\item{selected_covariate:}{The environmental covariate you'd like to run the model for. Can be
either \code{wy_type} (water year type), \code{max_flow_std} (maximum flow), \code{gdd_std} (growing degree days),
or \code{null_covar} (no environmental covariate).}
}
\value{
a list containing data elements required to run the STAN passage to spawner model
}
\description{
This function prepares data for running P2S on a single stream and covariate. It is called in \code{fit_passage_to_spawner_model()}.
}
\seealso{
Other Prepare Model Inputs:
\code{\link[=prepare_abundance_inputs]{prepare_abundance_inputs()}},
\code{\link[=prepare_inseason_inputs]{prepare_inseason_inputs()}},
\code{\link[=prepare_pCap_inputs]{prepare_pCap_inputs()}},
\code{\link[=prepare_stock_recruit_inputs]{prepare_stock_recruit_inputs()}},
\code{\link[=prepare_survival_inputs]{prepare_survival_inputs()}}
}
\concept{Prepare Model Inputs}
