% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{prepare_survival_inputs}
\alias{prepare_survival_inputs}
\title{Prepare inputs for survival model}
\usage{
prepare_survival_inputs(number_of_water_year_types = NULL, effect = NULL)
}
\arguments{
\item{number_of_water_year_types}{either \code{2}, \code{3}, or leave empty. If empty, the function will call a
version of the survival with no covariate effect.}

\item{effect}{one of \code{no_biological_effect}, \code{fork_length_effect}, \code{weight_effect}, or \code{condition}.
If supplying an effect, you must supply an argument for \code{number_of_water_year_types}}
}
\value{
a named list containing the inputs to the \code{run_survival_model()} function:
\itemize{
\item \strong{data_inputs} a named list of data inputs for passing to the STAN model.
\item \strong{inits} a list of initial values for passing to the STAN model.
\item \strong{parameters} a vector of parameter names to save from calling the STAN model.
}
}
\description{
Prepare inputs for survival model
}
\details{
This function prepares the data, initial values, and parameter list needed to call the
survival STAN model. It will prepare these data based on what model you want to run.
}
\seealso{
Other Prepare Model Inputs:
\code{\link[=prepare_P2S_inputs]{prepare_P2S_inputs()}},
\code{\link[=prepare_abundance_inputs]{prepare_abundance_inputs()}},
\code{\link[=prepare_inseason_inputs]{prepare_inseason_inputs()}},
\code{\link[=prepare_pCap_inputs]{prepare_pCap_inputs()}},
\code{\link[=prepare_stock_recruit_inputs]{prepare_stock_recruit_inputs()}}
}
\concept{Prepare Model Inputs}
