% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_season_model.R
\name{prepare_within_season_inputs}
\alias{prepare_within_season_inputs}
\title{Prepare inputs for within season model}
\usage{
prepare_within_season_inputs(con, stream, site)
}
\arguments{
\item{con}{A valid connection to the model run database.}

\item{stream}{The stream for which you want to fit the model.}

\item{site}{The site for which you want to fit the model.}
}
\value{
a list of inputs TBD
}
\description{
Prepare inputs for within season model
}
\details{
This function prepares data for input into a within season model.
}
