% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_model.R
\name{read_plad_csv}
\alias{read_plad_csv}
\title{Read in PLAD csv results}
\usage{
read_plad_csv(site)
}
\description{
helper function to read in PLAD results as currently provided in a .csv format.
}
\keyword{internal}
