% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_multiple_bt_spas_x}
\alias{run_multiple_bt_spas_x}
\title{Call BT-SPAS-X (WinBUGS) model on all site/run years}
\usage{
run_multiple_bt_spas_x(
  bt_spas_x_bayes_params,
  weekly_juvenile_abundance_catch_data,
  weekly_juvenile_abundance_efficiency_data,
  sites_to_run = NULL,
  effort_adjust,
  bugs_directory,
  debug_mode,
  no_cut = c(F, T)
)
}
\arguments{
\item{weekly_juvenile_abundance_catch_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_catch_data}}

\item{weekly_juvenile_abundance_efficiency_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_efficiency_data}}

\item{sites_to_run}{a subset of site/run year/lifestage combinations in tibble format to run}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort}

\item{bugs_directory}{where the \code{WinBUGS.exe} file can be found. Needs to end in \verb{/WinBUGS}}

\item{debug_mode}{whether you want to run \code{bugs} in debug mode.}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}. Can use \code{SRJPEmodel::bt_spas_x_bayes_params}.}
}
\value{
a list containing the results for every unique \code{site} and \verb{run year} combination
in the input data.
}
\description{
Call BT-SPAS-X (WinBUGS) model on all site/run years
}
\details{
This calls \code{run_single_bt_spas_x()} on all sites/run year/lifestage combinations
present in the input data.
}
