% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{run_passage_to_spawner_model}
\alias{run_passage_to_spawner_model}
\title{Run Passage to Spawner (P2S) STAN Model}
\usage{
run_passage_to_spawner_model(data_inputs)
}
\arguments{
\item{data_inputs}{A list object containing the following variables:
\itemize{
\item \strong{N} number of years
\item \strong{observed_passage} upstream passage for the associated years
\item \strong{observed_spawners} observed spawner counts for the associated years. Redd counts for Battle and Clear Creeks.
\item \strong{percent_female} Assumed sex ratio for the population and used to convert spawner counts to total population.
\code{0.5} for redds, \code{1} for other spawner survey types.
\item \strong{environmental_covar} the environmental covariate you want to run the model with.
\item \strong{ss_total} total sum of squares.
\item \strong{average_upstream_passage} average upstream passage across all years, used to forecast.
}}
}
\value{
a STANfit model object (see \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{details})
}
\description{
This functions take in a list of data required to run the Passage to Spawner STAN model. The function
will return the full fitted \code{stanfit} object.
See \code{vignette("passage_to_spawner_submodel.Rmd", package = "SRJPEmodel")} for more details.
}
\seealso{
Other passage_to_spawner: 
\code{\link{compare_P2S_model_covariates}()},
\code{\link{extract_P2S_estimates}()}
}
\concept{passage_to_spawner}
