% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_single_bt_spas_x}
\alias{run_single_bt_spas_x}
\title{Call BT-SPAS-X (WinBUGS) on a single site/run year combination}
\usage{
run_single_bt_spas_x(
  bt_spas_x_bayes_params,
  weekly_juvenile_abundance_catch_data,
  weekly_juvenile_abundance_efficiency_data,
  site,
  run_year,
  lifestage,
  effort_adjust = c(T, F),
  bugs_directory,
  debug_mode,
  no_cut = c(F, T)
)
}
\arguments{
\item{weekly_juvenile_abundance_catch_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_catch_data}}

\item{weekly_juvenile_abundance_efficiency_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_efficiency_data}}

\item{site}{site for which you want to fit the model}

\item{run_year}{run year for which you want to fit the model}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort}

\item{bugs_directory}{where the \code{WinBUGS.exe} file can be found. Needs to end in \verb{/WinBUGS}}

\item{debug_mode}{whether you want to run \code{bugs} in debug mode.}

\item{no_cut}{whether to run the model using the \code{cut()} function.}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}. Can use \code{SRJPEmodel::bt_spas_x_bayes_params}.}

\item{lifetage}{the lifestage for which you want tor run the model. One of \code{yearling}, \code{fry}, and \code{smolt}.}
}
\value{
a list:
\itemize{
\item \strong{results} model results - see \code{?bt_spas_x_bugs()} for details.
\item \strong{site} the site used to fit the model
\item \strong{run_year} the run year used to fit the model
\item \strong{weeks_fit} the weeks with catch used to fit the model (for analysis and plotting)
\item \strong{knots_output} knot positions for the weekly abundance spline curve
}
}
\description{
Call BT-SPAS-X (WinBUGS) on a single site/run year combination
}
\details{
This function is called within \code{run_bt_spas_x()} or can be run by itself
on a single site/run year combination.
}
