% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_single_bt_spas_x_stan}
\alias{run_single_bt_spas_x_stan}
\title{Call BT-SPAS-X using STAN}
\usage{
run_single_bt_spas_x_stan(
  bt_spas_x_bayes_params,
  weekly_juvenile_abundance_catch_data,
  weekly_juvenile_abundance_efficiency_data,
  site,
  run_year,
  lifestage,
  effort_adjust = c(T, F)
)
}
\arguments{
\item{weekly_juvenile_abundance_catch_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_catch_data}}

\item{weekly_juvenile_abundance_efficiency_data}{data frame containing weekly RST catch data. See
\code{?SRJPEdata::weekly_juvenile_abundance_efficiency_data}}

\item{site}{site for which you want to fit the model}

\item{run_year}{run year for which you want to fit the model}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}. Can use \code{SRJPEmodel::bt_spas_x_bayes_params}.}

\item{lifetage}{the lifestage for which you want tor run the model. One of \code{yearling}, \code{fry}, and \code{smolt}.}
}
\description{
Call BT-SPAS-X using STAN
}
\details{
This is a draft version of the function that calls STAN instead of WinBUGS.
}
