% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_model.R
\name{sample_composite_posterior}
\alias{sample_composite_posterior}
\title{Sample inseason model posterior}
\usage{
sample_composite_posterior(
  size_class_arg,
  reach_arg,
  week_fit_arg,
  weekly_replicates_arg
)
}
\description{
Uses weights to set replicates and then samples based on those
replicates for use in calculating / comparing weighted survival
}
\keyword{internal}
