% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{setup_azure_blob_backend}
\alias{setup_azure_blob_backend}
\title{Azure Blob Backend}
\usage{
setup_azure_blob_backend(
  storage_account,
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY")
)
}
\arguments{
\item{storage_account}{Azure storage account name.}

\item{access_key}{Azure storage access key.}
}
\value{
An \code{AzureStor} storage endpoint object.
}
\description{
Creates an \code{AzureStor} storage endpoint object.
}
\keyword{internal}
