% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_model_code}
\alias{survival_model_code}
\title{Passage to Spawner model code}
\format{
An object of class \code{character} of length 1.
}
\usage{
survival_model_code
}
\description{
A character string containing the STAN model code for the juvenile outmigration survival submodel.
}
\keyword{datasets}
