% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{bt_spas_x_bugs}
\alias{bt_spas_x_bugs}
\title{Execute BT-SPAS-X in WinBUGs}
\usage{
bt_spas_x_bugs(
  data,
  inits,
  parameters,
  model_name,
  bt_spas_x_bayes_params,
  number_mcmc,
  bugs_directory,
  debug_mode,
  no_cut
)
}
\arguments{
\item{data}{a list containing the following elements:
\itemize{
\item \strong{Nmr} number of unique mark-recapture experiments performed across tributaries, years and weeks
\item \strong{Ntribs} number of tributaries to use for the pCap component of the model
\item \strong{Nstrata} number of weeks with catch data
\item \strong{Uwc_ind} number of weeks in catch data with associated pCap and flow data
\item \strong{Nwomr} number of weeks in catch data without associated pCap and flow data
\item \strong{Nstrata_wc} number of weeks in catch data with catch data (RST fished)
\item \strong{indices_site_pCap} tributary index (of all possible tributaries) to use to predict efficiency for missing strata. Note length=0 if selected tributary is not part of trib set that has mark-recap data. In this case model that samples from trib hyper will be called.
\item \strong{ind_trib} indices (1:Ntribs) assigned to the mark-recapture experiment table for use in BUGs
\item \strong{ind_pCap} indices of weeks in mark-recapture table for U strata being estimated
\item \strong{Uind_woMR} indices of weeks in catch data without associated pCap and flow data
\item \strong{Uind_wMR} indices of weeks in catch data with associated pCap and flow data
\item \strong{Uwc_ind} indices of weeks in catch data with catch data
\item \strong{Releases} number of fish released for each mark-recapture experiment
\item \strong{Recaptures} number of fish recaptured for each mark-recapture experiment
\item \strong{u} weekly abundance
\item \strong{mr_flow} standardized flow, averaged over recapture days (< 1 week)
\item \strong{catch_flow} standardized flow, averaged by week
\item \strong{K} number of columns in the bspline basis matrix
\item \strong{ZP} The b_spline_matrix (bspline basis matrix. One row for each data point (1:Nstrata), and one column for each term in the cubic polynomial function (4) + number of knots)
\item \strong{lgN_max} maxmimum possible value for log N across strata
}}

\item{inits}{a list containing initial values for the following parameters:
\itemize{
\item \strong{trib_mu.P} mean of hyper-distribution for site-effect
\item \strong{b0_pCap} site effect on trap efficiency for each site
\item \strong{flow_mu.P} mean of hyper-distribution for flow effect
\item \strong{b_flow} flow effect on trap efficiency for each site
\item \strong{trib_tau.P} used to estimate standard deviation of hyper-distribution for site effect
\item \strong{flow_tau.P} used to estimate standard deviation of hyper-distribution for flow effect
\item \strong{pro_tau.P} used to estimate standard deviation of zero-centered normal distribution for unexplained error
\item \strong{b_sp} basis function of each spline node
\item \strong{lg_N} predicted weekly abundance
}}

\item{parameters}{a list of parameters to be estimated in the model:
\itemize{
\item \strong{trib_mu.P} mean of hyper-distribution for site-effect
\item \strong{trib_sd.P} standard deviation of hyper-distribution for site effect
\item \strong{flow_mu.P} mean of hyper-distribution for flow effect
\item \strong{flow_sd.P} standard deviation of hyper-distribution for flow effect
\item \strong{pro_sd.P} standard deviation of zero-centered normal distribution for unexplained error
\item \strong{b0_pCap} site effect on trap efficiency for each site
\item \strong{b_flow} flow effect on trap efficiency for each site
\item \strong{pCap_U} weekly trap efficiency (capture probability)
\item \strong{N} weekly juvenile abundance
\item \strong{Ntot} total juvenile abundance for the year
\item \strong{sd.N} standard deviation controlling flexibility of spline weekly abundance curve
\item \strong{sd.Ne} standard deviation controlling extent of non-spline variation in weekly abundance
}}

\item{model_name}{model to be called based on number of efficiency trials available for the selected site. Either
\itemize{
\item \strong{all_mark_recap.bug} all weeks with catch have corresponding efficiency trials
\item \strong{missing_mark_recap.bug} some weeks with catch have corresponding efficiency trials
\item \strong{no_mark_recap_no_trib.bug} the selected site has no efficiency data at all
\item \strong{no_mark_recap.bug} no weeks with catch have corresponding efficiency trials
}}

\item{bugs_directory}{a filepath indicating where to find the \verb{WinBUGS14/} file. This needs
to be in a character format ending with \verb{/WinBUGS14}.}

\item{debug_mode}{whether you want to run \code{bugs} in debug mode.}

\item{no_cut}{whether to run the model using the \code{cut()} function.}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}.}
}
\value{
if running on a Mac operating system, returns a list of all inputs formatted
to pass to \code{bugs}. If running on a PC or another operating system capable of running WinBUGS,
returns a nested list containing the following elements:
\itemize{
\item \strong{model_results} the \code{BUGs} object from fitting the model
\item \strong{model_called} the model called on the databugs_
\item \strong{data_inputs} the data passed to the model
\item \strong{init_inputs} the initial values passed to the model
}
}
\description{
Execute BT-SPAS-X in WinBUGs
}
\details{
This function is called within \code{run_single_bt_spas_x()} and calls the WinBUGS code
}
