% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bt_spas_x_model_code}
\alias{bt_spas_x_model_code}
\title{BT-SPAS-X STAN model code}
\format{
A named list with 4 elements:
\itemize{
\item \code{pCap_all}: pCap model code
\item \code{abundance}: abundance model code (STAN)
\item \code{abundance_BUGS}: abundance model code (BUGS)
}
}
\usage{
bt_spas_x_model_code
}
\description{
A nested named list containing model code for versions of the bt-spas-x models.
Model code written in WinBUGS has two versions: "with" and "without" cut functions. The
STAN
}
\keyword{datasets}
