% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{extract_abundance_estimates}
\alias{extract_abundance_estimates}
\title{Extract parameter estimates from abundance BUGS model object}
\usage{
extract_abundance_estimates(site, run_year, abundance_inputs, model_object)
}
\arguments{
\item{site}{the site fit.}

\item{run_year}{the run year fit.}

\item{abundance_inputs}{a list of inputs for the abundance BUGS model, generated by running \code{prepare_abundance_inputs()}}

\item{model_object}{the BUGS object produced by running \code{fit_abundance_model_BUGS()}.}
}
\value{
A table with the format:
\itemize{
\item \strong{model_name}
\item \strong{site}
\item \strong{run_year}
\item \strong{week_fit}
\item \strong{location_fit}
\item \strong{parameter}
\item \strong{statistic}
\item \strong{value}
\item \strong{srjpedata_version}
}
}
\description{
Extract parameter estimates from abundance BUGS model object
}
\details{
This function extracts parameter estimates from the abundance BUGS model and returns them in a tidy data frame format.
}
