% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{extract_bt_spas_x_results}
\alias{extract_bt_spas_x_results}
\title{Extract results from BT-SPAS-X bugs object}
\usage{
extract_bt_spas_x_results(model_fit_object)
}
\arguments{
\item{model_fit_object}{the object produced by running \code{run_single_bt_spas_x()}
for a given site and run year.}
}
\value{
a named list with the following elements:
\itemize{
\item \strong{posterior_output} the posterior simulations produced by the \code{BUGs} object
\item \strong{summary_output} the summary produced by the \code{BUGs} object
\item \strong{dic_output} DIC for the \code{BUGs} object
\item \strong{knots_output} the knot positions produced for the weekly abundance spline curve
}
}
\description{
Extract results from BT-SPAS-X bugs object
}
\details{
This function is called within \code{run_single_bt_spas_x()} and prepares spline parameters
to pass to the data list. These are helpful for model diagnostics for a given model run.
}
