% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{extract_bt_spas_x_results}
\alias{extract_bt_spas_x_results}
\title{Extract results from BT-SPAS-X bugs object}
\usage{
extract_bt_spas_x_results(model_result_object)
}
\arguments{
\item{model_result_object}{TODO}
}
\value{
TODO
}
\description{
Extract results from BT-SPAS-X bugs object
}
\details{
TODO
}
