% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_season_model.R
\name{extract_inseason_estimates}
\alias{extract_inseason_estimates}
\title{Extract parameter estimates from in-season model}
\usage{
extract_inseason_estimates(inputs, fit)
}
\arguments{
\item{inputs}{a list produced by running \code{prepare_inseason_inputs()}}

\item{fit}{a stanfit object produced by running \code{fit_inseason_model()}}
}
\value{
a table
}
\description{
This produces a formatted data frame of parameter estimates for the inseason STAN model.
}
