% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{extract_pCap_estimates}
\alias{extract_pCap_estimates}
\title{Extract parameter estimates from the pCap BUGS model object}
\usage{
extract_pCap_estimates(model_object, pCap_inputs)
}
\arguments{
\item{model_object}{the BUGS object produced by running \code{fit_pCap_model_BUGS()}.}

\item{pCap_inputs}{a list of inputs for the abundance BUGS model, generated by running \code{prepare_pCap_inputs()}}
}
\value{
A table with the format:
\itemize{
\item \strong{model_name}
\item \strong{site}
\item \strong{run_year}
\item \strong{week_fit}
\item \strong{location_fit}
\item \strong{parameter}
\item \strong{statistic}
\item \strong{value}
\item \strong{srjpedata_version}
}
}
\description{
Extract parameter estimates from the pCap BUGS model object
}
\details{
This function extracts parameter estimates from the pCap BUGS model and returns them in a tidy data frame format.
}
