% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{fit_abundance_model_BUGS}
\alias{fit_abundance_model_BUGS}
\title{Fit abundance model in BUGS.}
\usage{
fit_abundance_model_BUGS(
  abundance_inputs,
  lt_pCap_Us,
  bugs_model_file,
  bugs_directory
)
}
\arguments{
\item{abundance_inputs}{the object produced by \code{prepare_abundance_inputs()}}

\item{lt_pCap_Us}{the object produced by \code{generate_lt_pCap_Us()} containing values of
length \code{Nstrata} for \code{lt_pCap_mu} and \code{lt_pCap_sd}.}

\item{bugs_model_file}{the filepath pointing to where your BUGS abundance model file is}

\item{bugs_directory}{the filepath pointing to where your WinBUGS14/ directory is}
}
\value{
a BUGS object.
}
\description{
Fit abundance model in BUGS.
}
\details{
This function runs the BUGS abundance model.
}
