% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bt_spas_inputs.R
\name{fit_abundance_model_BUGS}
\alias{fit_abundance_model_BUGS}
\title{Run BUGS abundance model.}
\usage{
fit_abundance_model_BUGS(abundance_inputs, lt_pCap_Us, bugs_model_file)
}
\arguments{
\item{abundance_inputs}{the object produced by \code{prepare_abundance_inputs()}}

\item{lt_pCap_Us}{the object produced by \code{generate_lt_pCap_Us()}}
}
\value{
a BUGS object.
}
\description{
Run BUGS abundance model.
}
\details{
This function runs the BUGS abundance model.
}
