% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_season_model.R
\name{fit_inseason_model}
\alias{fit_inseason_model}
\title{Fit In-Season STAN model}
\usage{
fit_inseason_model(inputs)
}
\arguments{
\item{inputs}{a list generated by running \code{prepare_in_season_inputs()}}
}
\value{
a STANfit model object (see \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{details})
}
\description{
This functions take in a list of data required to run the in-season STAN model. The function
will return the full fitted \code{stanfit} object.
See \code{vignette("passage_to_spawner_submodel.Rmd", package = "SRJPEmodel")} for more details.
}
