% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{fit_pCap_model}
\alias{fit_pCap_model}
\title{Fit pCap model}
\usage{
fit_pCap_model(input)
}
\arguments{
\item{input}{A list containing the inputs for the pCap model.}
}
\value{
a named list with the required elements for that model run.
}
\description{
Fit pCap model
}
\details{
This function prepares the data list for the abundance STAN model based on what the model name is.
}
\keyword{internal}
