% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{generate_diagnostic_plot}
\alias{generate_diagnostic_plot}
\title{model_ids <- get_most_recent_model_results(con) |>
distinct(model_run_id) |>
pull(model_run_id)
if (model_component == "model_fit"){
model_urls <- tbl(con, "model_run") |>
filter(id \%in\% model_ids) |>
pull(blob_fit_storage_url)
}else if (model_component == "model_input"){
model_urls <- tbl(con, "model_run") |>
filter(id \%in\% model_ids) |>
pull(blob_input_storage_url)
} else {
model_urls <- tbl(con, "model_run") |>
filter(id \%in\% model_ids) |>
pull(blob_plot_storage_url)
}}
\usage{
generate_diagnostic_plot(inputs, fit)
}
\arguments{
\item{inputs}{inputs for the fit model, created by running the appropriate \code{prepare_inputs} function.}

\item{fit}{the model fit}

\item{model_results_name}{The model type name associated with options in the database: either
\code{all_mark_recap}, \code{no_mark_recap}, \code{missing_mark_recap}, \code{no_mark_recap_no_trib},
\code{pcap_all}, \code{pcap_mainstem}, \code{p2s}, or \code{stock_recruit}.}
}
\value{
Generates a diagnostic posterior predictive plot for a model fit.
}
\description{

}
\details{

}
