% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{generate_diagnostic_plot}
\alias{generate_diagnostic_plot}
\title{Plot model diagnostics}
\usage{
generate_diagnostic_plot(model_fit_summary_object, model_language)
}
\arguments{
\item{model_fit_summary_object}{The model summary object produced by a model fit}
}
\value{
A two-panel plot showing abundance estimates and weekly efficiency estimates compared
to lincoln-peterson estimates and raw count and efficiency trial data.
}
\description{
Plot model diagnostics
}
