% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{generate_diagnostic_plot_juv}
\alias{generate_diagnostic_plot_juv}
\title{BT SPAS X diagnostic plots}
\usage{
generate_diagnostic_plot_juv(inputs, results_df)
}
\arguments{
\item{inputs}{a list of inputs generated by running \code{prepare_abundance_inputs()}}

\item{results_df}{a table of parameter estimates generated by running \code{extract_abundance_estimates()} or
by pulling and filtering from the database using \code{get_most_recent_model_results()}.}
}
\value{
A plot.
}
\description{
BT SPAS X diagnostic plots
}
\details{
This function produces a plot with data and results of fitting the pCap and abundance models for
a given site and run year.
}
