% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{generate_diagnostic_plot_juv}
\alias{generate_diagnostic_plot_juv}
\title{BT SPAS X diagnostic plots}
\usage{
generate_diagnostic_plot_juv(site_arg, run_year_arg, abundance_model_fit_table)
}
\arguments{
\item{site_arg}{The site being fit}

\item{run_year_arg}{The run year being fit}

\item{abundance_model_fit_table}{A table that is produced from running \code{extract_abundance_estimates()}.}
}
\value{
A plot.
}
\description{
BT SPAS X diagnostic plots
}
\details{
This function produces a plot with data and results of fitting the pCap and abundance models for
a given site and run year.
}
