% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{generate_lt_pCap_Us}
\alias{generate_lt_pCap_Us}
\title{Generate lt_pCap_U values from the pCap model object}
\usage{
generate_lt_pCap_Us(abundance_inputs, pCap_model_object)
}
\arguments{
\item{pCap_model_object}{A STANfit object with the output of the pCap hierarchical model.}

\item{pCap_inputs}{A list containing the inputs for the pCap model including \code{model_name}, \code{Nstrata},
\code{catch_flow}, \code{use_trib}, \code{Ntribs}, \code{Nmr}, \code{Nwmr}, \code{Nwomr}, \code{Uind_wMR}, \code{Uind_woMR},
\code{ind_pCap}.}
}
\value{
a named list with the required elements for that model run.
}
\description{
Generate lt_pCap_U values from the pCap model object
}
\details{
This function prepares the data list for the abundance STAN model based on what the model name is.
}
