% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{generate_results_plot_sr}
\alias{generate_results_plot_sr}
\title{Stock-recruit results plots}
\usage{
generate_results_plot_sr(sr_inputs, con)
}
\arguments{
\item{sr_inputs}{inputs for the fit model, created by running \code{prepare_stock_recruit_inputs()}}

\item{con}{a connection to the database.}
}
\value{
A plot.
}
\description{
Stock-recruit results plots
}
\details{
This function produces a plot with observed and predicted recruits
for simple comparison.
}
