% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_spas_x_refactor.R
\name{get_bt_spas_x_data_list}
\alias{get_bt_spas_x_data_list}
\title{Prepare Data Object for BT-SPAS-X WinBUGs call}
\usage{
get_bt_spas_x_data_list(model_name, full_data_list)
}
\value{
a named list with the required elements for that model run.
}
\description{
Prepare Data Object for BT-SPAS-X WinBUGs call
}
\details{
This function is called within \code{run_single_bt_spas_x()} and prepares the data lists
based on which model will be run.
}
