% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_hyper_distribution_parameter_estimates}
\alias{get_hyper_distribution_parameter_estimates}
\title{Extract hyper-distribution parameters from BT-SPAS-X}
\usage{
get_hyper_distribution_parameter_estimates(model_fit_object)
}
\description{
This function can be run on objects produced by \code{run_bt_spas_x()}.
It extracts estimates for \code{trib_mu}, \code{trib_sd}, \code{flow_mu}, \code{flow_sd}, and \code{pro_sd}.
}
\keyword{internal}
