% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_most_recent_model_objects}
\alias{get_most_recent_model_objects}
\title{Get Most Recent Model Objects}
\usage{
get_most_recent_model_objects(
  con,
  model_component = "model_fit",
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY"),
  model_name = NULL,
  stream = NULL
)
}
\arguments{
\item{con}{A connection object to the database.}

\item{model_component}{A choice of \code{model_fit}, \code{model_input} or \code{model_plot} to pull.}

\item{model_name}{Can be left empty. If provided, must be one of \code{pcap_all}, \code{bt_spas_x},
\code{pcap_mainstem}, \code{p2s}, \code{inseason}, or \code{stock_recruit}}

\item{stream}{Can be left empty. If provided, must be one of \verb{battle creek},
\verb{clear creek}, \verb{deer creek}, \verb{mill creek}, \verb{feather river}, \verb{butte creek},
\verb{sacramento river}, or \verb{yuba river}.}
}
\value{
A the most recent model fit objects, inputs, or plots. The format will be a named list,
where each element is named by the \code{model_run_id}, \code{model_name}, \code{site}, \code{stream}, and \code{year}.
}
\description{
This function retrieves the most recent model objects for each model name, site, year, stream, etc. You can
specify model_object (the full fit object, either \code{BUGS} or \code{stanfit}, the model plot (a posterior predictive check plot), or
the inputs used to fit the associated model object.
}
