% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_most_recent_model_objects}
\alias{get_most_recent_model_objects}
\title{Get Most Recent Model Run Objects}
\usage{
get_most_recent_model_objects(
  con,
  model_name_filter = NULL,
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY")
)
}
\arguments{
\item{con}{A connection object to the database.}

\item{model_name_filter}{An optional argument specifying which kind of model you want to pull. This can be either
\code{bt_spas_x}, \code{pcap_mainstem}, \code{pcap_all}, \code{p2s}, or \code{stock_recruit})}
}
\value{
A list of model objects.
}
\description{
This function retrieves the most recent model run objects for each model name, site, year, stream, etc.
}
