% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_most_recent_model_objects}
\alias{get_most_recent_model_objects}
\title{Get Most Recent Model Run Objects}
\usage{
get_most_recent_model_objects(
  con,
  model_component = "model_fit",
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY")
)
}
\arguments{
\item{con}{A connection object to the database.}

\item{model_component}{A choice of model_fit, model_input or model_plot to pull.}
}
\value{
A the most recent model objects.
}
\description{
This function retrieves the most recent model run objects for each model name, site, year, stream, etc.
}
