% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Extract summary in table}
\usage{
get_summary_table(
  model_fit_object,
  site,
  run_year,
  lifestage,
  weeks_fit,
  sites_fit,
  model_name
)
}
\description{
TODO
}
\keyword{internal}
