% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_total_juvenile_abundance}
\alias{get_total_juvenile_abundance}
\title{Extract total juvenile abundance estimates}
\usage{
get_total_juvenile_abundance(juvenile_model_fits_summary)
}
\description{
TODO
}
\keyword{internal}
