% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_total_juvenile_abundance}
\alias{get_total_juvenile_abundance}
\title{Extract total juvenile abundance estimates}
\usage{
get_total_juvenile_abundance(model_fit_object)
}
\description{
This function can be run on objects produced by \code{run_bt_spas_x()}.
It extracts the estimate for \code{Ntot}, or total abundance.
}
\keyword{internal}
