% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_weekly_juvenile_abundance}
\alias{get_weekly_juvenile_abundance}
\title{Extract weekly juvenile abundance}
\usage{
get_weekly_juvenile_abundance(model_fit_object)
}
\description{
This function can be run on objects produced by \code{run_bt_spas_x()}.
It extracts estimates for \code{N}, or abundance, by week.
}
\keyword{internal}
