% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{get_weekly_trap_efficiency}
\alias{get_weekly_trap_efficiency}
\title{Extract weekly trap efficiency from BT-SPAS-X object}
\usage{
get_weekly_trap_efficiency(model_fit_object)
}
\description{
This function can be run on objects produced by \code{run_bt_spas_x()}.
It extracts estimates for \code{pCap} by week, aka weekly trap efficiency.
}
\keyword{internal}
