% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{model_pin_board}
\alias{model_pin_board}
\title{Model Data Pin Board}
\usage{
model_pin_board(storage_account, container, ...)
}
\arguments{
\item{storage_account}{The name of the Azure Storage account.}

\item{container}{The name of the blob container within the storage account.}

\item{...}{Additional arguments passed to \code{setup_azure_blob_backend()}.}
}
\value{
A pins board object connected to the specified Azure Blob Storage container.
}
\description{
Creates a pin board for model data storage in Azure Blob Storage.
}
\examples{
\dontrun{
board <- model_pin_board("my_storage_account", "my_container")
}

}
\seealso{
\code{\link[AzureStor]{blob_container}}
\code{\link[pins]{board_azure}}
}
