% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{plot_juv_data}
\alias{plot_juv_data}
\title{BT SPAS X raw data plots}
\usage{
plot_juv_data(site, run_year)
}
\arguments{
\item{site}{The site being fit}

\item{run_year}{The run year being fit}
}
\value{
A plot
}
\description{
BT SPAS X raw data plots
}
\details{
This function produces a plot with data for a site and run year.
}
