% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{prepare_abundance_inputs}
\alias{prepare_abundance_inputs}
\title{Prepare inputs for abundance model}
\usage{
prepare_abundance_inputs(
  site,
  run_year,
  pCap_model_type = c("one_site", "one_site_skew", "all_sites"),
  min_pCap = 5e-04,
  min_pCap_mult = 1,
  pCap_model_object,
  input_catch_data = NULL,
  input_efficiency_data = NULL
)
}
\arguments{
\item{site}{site for which you want to fit the model}

\item{run_year}{run year for which you want to fit the model}

\item{pCap_model_type}{One of \code{all_sites}, \code{one_site_skew}, or \code{one_site}. Describes your input
for parameter \code{pCap_model_object}}

\item{min_pCap}{a parameter that sets the minimum pCap value for setting maximum and initial
values for abundance (N). Can be tweaked by min_pCap_mult. Default value is \code{0.0005}}

\item{min_pCap_mult}{a parameter that lets you tweak the min pCap value for setting
maximum and initial values for abundance (N). Gets multiplied by min_pCap. Default value is \code{1.0}}

\item{pCap_model_object}{the STAN model file produced by running \code{prepare_pCap_inputs()}
#' @param input_catch_data Optional argument for weekly catch data.
Defaults to \verb{SRJPEdata::SRJPEdata::weekly_juvenile_abundance_catch_data}. If
passed in, structure of data frame must match that of the default.}

\item{input_catch_data}{Optional argument for weekly catch data.
Defaults to \code{SRJPEdata::weekly_juvenile_abundance_catch_data}. If
passed in, structure of data frame must match that of the default.}

\item{input_efficiency_data}{Optional argument for weekly efficiency data.
Defaults to \verb{SRJPEdata::SRJPEdata::weekly_juvenile_abundance_efficiency_data}. If
passed in, structure of data frame must match that of the default.}
}
\value{
a list:
\itemize{
\item \strong{abundance_inputs} a list of data and inits for input into the abundance model.
\item \strong{model_name} The version of the pCap logic to use for generating lt_pCap_U values. Either
\code{all_mark_recap}, \code{missing_mark_recap}, \code{no_mark_recap}, or \code{no_mark_recap_no_trib}.
\item \strong{site} The site run.
\item \strong{run_year} The run year.
\item \strong{weeks_fit} The weeks fit for the abundance model
\item \strong{weeks_date} Associated dates for the weeks fit for the abundance model.
\item \strong{lt_pCap_Us} A list containing output of \code{generate_lt_pCap_Us}: containing values of
length \code{Nstrata} for \code{lt_pCap_mu} and \code{lt_pCap_sd}.
}
}
\description{
Prepare inputs for abundance model
}
\details{
This function prepares data for input into an abundance model.
}
