% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bt_spas_inputs.R
\name{prepare_abundance_inputs}
\alias{prepare_abundance_inputs}
\title{Prepare inputs for abundance model}
\usage{
prepare_abundance_inputs(
  input_catch_data,
  input_efficiency_data,
  site,
  run_year,
  effort_adjust = c(T, F)
)
}
\arguments{
\item{site}{site for which you want to fit the model}

\item{run_year}{run year for which you want to fit the model}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort.}
}
\value{
a list:
\itemize{
\item \strong{abundance_inputs} a list of data and inits for input into the abundance model.
\item \strong{model_name} The version of the pCap logic to use for generating lt_pCap_U values. Either
\code{all_mark_recap}, \code{missing_mark_recap}, \code{no_mark_recap}, or \code{no_mark_recap_no_trib}.
\item \strong{weeks_fit} The weeks fit for the abundance model
\item \strong{weeks_date} Associated dates for the weeks fit for the abundance model.
}
}
\description{
Prepare inputs for abundance model
}
\details{
This function prepares data for input into an abundance model.
}
