% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{prepare_pCap_inputs}
\alias{prepare_pCap_inputs}
\title{Prepare inputs for pCap STAN model}
\usage{
prepare_pCap_inputs(
  model_type = c("all_sites", "one_site"),
  skew = NULL,
  site_selection = NULL,
  exclude_from_hyper = NULL,
  input_catch_data = NULL,
  input_efficiency_data = NULL
)
}
\arguments{
\item{model_type}{either \code{one_site} or \code{all_sites}}

\item{skew}{either \code{TRUE} or \code{FALSE}. Only applies for \code{one_site} models.}

\item{site_selection}{if fitting the \code{one_site} model, the tributary you want to fit.}

\item{exclude_from_hyper}{a vector of sites you want to drop from being estimated in the hyper distribution
parameters \code{b0_pCap} and \code{b_flow}. Default is NULL, meaning no sites are dropped. Must be site names
in \code{unique(SRJPEdata::weekly_juvenile_abundance_efficiency_data$site)}}

\item{input_catch_data}{Optional argument for weekly catch data.
Defaults to \code{SRJPEdata::weekly_juvenile_abundance_catch_data}. If
passed in, structure of data frame must match that of the default.}

\item{input_efficiency_data}{Optional argument for weekly efficiency data.
Defaults to \verb{SRJPEdata::SRJPEdata::weekly_juvenile_abundance_efficiency_data}. If
passed in, structure of data frame must match that of the default.}
}
\value{
a list:
\itemize{
\item \strong{pCap_inputs} a list of data and inits for input into the pCap model.
\item \strong{sites_fit} a list of site names associated with \code{ind_trib}.
}
}
\description{
Prepare inputs for pCap STAN model
}
\details{
This function prepares data for input into a pCap STAN model.
}
\seealso{
Other Prepare Model Inputs: 
\code{\link{prepare_P2S_inputs}()},
\code{\link{prepare_inseason_inputs}()},
\code{\link{prepare_stock_recruit_inputs}()},
\code{\link{prepare_survival_inputs}()}
}
\concept{Prepare Model Inputs}
