% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{prepare_pCap_inputs}
\alias{prepare_pCap_inputs}
\title{Prepare inputs for pCap STAN model}
\usage{
prepare_pCap_inputs(
  mainstem = c(FALSE, TRUE),
  mainstem_site = NULL,
  input_catch_data = NULL,
  input_efficiency_data = NULL
)
}
\arguments{
\item{mainstem}{Whether or not you want to evaluate efficiency trials for a mainstem site
(\verb{knights landing}, \code{tisdale}, and \verb{red bluff diversion dam}) or for a tributary site. If \code{FALSE},
the mark recapture dataset will be filtered to exclude those mainstem sites.}

\item{mainstem_site}{If you are fitting the model with \code{mainstem == TRUE}, you must supply a
mainstem site for which to prepare inputs. Can be either \verb{knights landing} or \code{tisdale}.}

\item{input_catch_data}{Optional argument for weekly catch data.
Defaults to \code{SRJPEdata::weekly_juvenile_abundance_catch_data}. If
passed in, structure of data frame must match that of the default.}

\item{input_efficiency_data}{Optional argument for weekly efficiency data.
Defaults to \verb{SRJPEdata::SRJPEdata::weekly_juvenile_abundance_efficiency_data}. If
passed in, structure of data frame must match that of the default.}
}
\value{
a list:
\itemize{
\item \strong{pCap_inputs} a list of data and inits for input into the pCap model.
\item \strong{sites_fit} a list of site names associated with \code{ind_trib}.
}
}
\description{
Prepare inputs for pCap STAN model
}
\details{
This function prepares data for input into a pCap STAN model.
}
