% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bt_spas_inputs.R
\name{prepare_pCap_inputs}
\alias{prepare_pCap_inputs}
\title{Prepare inputs for pCap STAN model}
\usage{
prepare_pCap_inputs(mainstem = c(FALSE, TRUE))
}
\arguments{
\item{mainstem}{Whether or not you want to evaluate efficiency trials for a mainstem site
(\verb{knights landing}, \code{tisdale}, and \verb{red bluff diversion dam}) or for a tributary site. If \code{FALSE},
the mark recapture dataset will be filtered to exclude those mainstem sites.}
}
\value{
a list:
\itemize{
\item \strong{pCap_inputs} a list of data and inits for input into the pCap model.
\item \strong{sites_fit} a list of site names associated with \code{ind_trib}.
}
}
\description{
Prepare inputs for pCap STAN model
}
\details{
This function prepares data for input into a pCap STAN model.
}
