% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{prepare_stock_recruit_inputs}
\alias{prepare_stock_recruit_inputs}
\title{Prepare inputs for stock-recruit model}
\usage{
prepare_stock_recruit_inputs(
  con,
  stream,
  adult_data_type,
  covariate,
  truncate_dataset = FALSE
)
}
\arguments{
\item{con}{A valid connection to the model run database.}

\item{stream}{The stream for which you want to fit the model.}

\item{adult_data_type}{the type of survey the adult data came from. Either \code{passage},
\code{redd}, \code{holding}, \code{broodstock_tag}, or \code{carcass}.}

\item{covariate}{The covariate you would like to use to fit the model. One of \code{rearing_max_flow}, \code{rearing_mean_flow},
\code{rearing_median_flow}, \code{rearing_min_flow}, \code{spawning_max_flow}, \code{spawning_mean_flow}, \code{spawning_median_flow}, \code{spawning_min_flow},
\code{spawning_above_13_temp_day}, \code{spawning_above_13_temp_week}, \code{spawning_gdd_spawn}, \code{spawning_weekly_max_temp_max},
\code{spawning_weekly_max_temp_mean}, \code{spawning_weekly_max_temp_median}, or \code{null}.}

\item{truncate_dataset}{either \code{TRUE} or \code{FALSE}. If \code{TRUE}, will filter the dataset to only include years where
there are data for ALL covariates. If \code{FALSE}, will filter the dataset to include years where there are data
for YOUR SELECTED covariate.}
}
\value{
a list of tables:
\itemize{
\item \strong{stock_recruit_table} juvenile abundance and adult abundance by brood year.
\item \strong{full_covariate_table} full table of covariates associated with the site, stream, and brood years in the stock recruit table
\item \strong{truncated_covariate_table} table of covariates limited to only those years where all covariates are available (for covariate comparison analyses)
}
}
\description{
Prepare inputs for stock-recruit model
}
\details{
This function prepares data for input into a stock recruit model.
}
\seealso{
Other Prepare Model Inputs: 
\code{\link{prepare_P2S_inputs}()},
\code{\link{prepare_abundance_inputs}()},
\code{\link{prepare_inseason_inputs}()},
\code{\link{prepare_pCap_inputs}()},
\code{\link{prepare_survival_inputs}()}
}
\concept{Prepare Model Inputs}
