% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_recruit.R
\name{prepare_stock_recruit_inputs}
\alias{prepare_stock_recruit_inputs}
\title{Prepare inputs for stock-recruit model}
\usage{
prepare_stock_recruit_inputs(
  con,
  year,
  stream,
  adult_data_type,
  covariate,
  truncate_dataset = FALSE
)
}
\arguments{
\item{con}{A valid connection to the model run database.}

\item{adult_data_type}{the type of survey the adult data came from. Either \code{upstream_estimate},
\code{redd_count}, \code{holding_count}, or \code{carcass_estimate}.}
}
\value{
a list of tables:
\itemize{
\item \strong{stock_recruit_table} juvenile abundance and adult abundance by brood year.
\item \strong{full_covariate_table} full table of covariates associated with the site, stream, and brood years in the stock recruit table
\item \strong{truncated_covariate_table} table of covariates limited to only those years where all covariates are available (for covariate comparison analyses)
}
}
\description{
Prepare inputs for stock-recruit model
}
\details{
This function prepares data for input into a stock recruit model.
}
