% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{prepare_survival_inputs}
\alias{prepare_survival_inputs}
\title{Prepare inputs for survival model}
\usage{
prepare_survival_inputs()
}
\value{
a named list containing the inputs to the \code{run_survival_model()} function:
\itemize{
\item \strong{data_inputs} a named list of data inputs for passing to the STAN model.
\item \strong{inits} a list of initial values for passing to the STAN model.
\item \strong{parameters} a vector of parameter names to save from calling the STAN model.
}
}
\description{
Prepare inputs for survival model
}
\details{
This function prepares the data, initial values, and parameter list needed to call the
survival STAN model. It will prepare these data based on what model you want to run. This
iteration of the model code runs the model using \code{max_flow} and \code{fork_length} as covariates. A future
version of the model code will allow for the full suite of covariates and biological effects.
}
\seealso{
Other Prepare Model Inputs: 
\code{\link{prepare_P2S_inputs}()},
\code{\link{prepare_inseason_inputs}()},
\code{\link{prepare_pCap_inputs}()},
\code{\link{prepare_stock_recruit_inputs}()}
}
\concept{Prepare Model Inputs}
