% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_model.R
\name{prepare_survival_inputs}
\alias{prepare_survival_inputs}
\title{Prepare inputs for survival model}
\usage{
prepare_survival_inputs(use_covariate, number_of_water_year_types, effect)
}
\arguments{
\item{use_covariate}{either \code{TRUE} or \code{FALSE}. If \code{TRUE}, will call a version of the survival
model that fits a covariate effect (water year type). If \code{FALSE}, it will call a version of the survival model
with no covariate effect.}

\item{number_of_water_year_types}{either \code{2} or \code{3}}

\item{effect}{one of \code{no_size_effect}, \code{size_effect}, \code{fork_length_effect}, \code{weight_effect}, \code{condition},}
}
\value{
TODO
}
\description{
Prepare inputs for survival model
}
\details{
Runs the survival model for both Sacramento releases and Feather/Butte releases.
}
