% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_abundance_workflow}
\alias{run_abundance_workflow}
\title{Run abundance BUGS workflow.}
\usage{
run_abundance_workflow(site, run_year, pCap, bugs_model_file, bugs_directory)
}
\value{
A table.
}
\description{
Run abundance BUGS workflow.
}
\details{
This function runs the abundance BUGS workflow.
}
