% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_bt_spas_x}
\alias{run_bt_spas_x}
\title{Call BT-SPAS-X model}
\usage{
run_bt_spas_x(
  bt_spas_x_bayes_params,
  bt_spas_x_input_data,
  site,
  run_year,
  effort_adjust,
  multi_run_mode,
  mainstem_version,
  special_priors_data,
  bugs_directory
)
}
\arguments{
\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}.}
}
\value{
TODO
}
\description{
Call BT-SPAS-X model
}
\details{
TODO
}
