% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_bt_spas_x_JPE_sites}
\alias{run_bt_spas_x_JPE_sites}
\title{Run all JPE sites.}
\usage{
run_bt_spas_x_JPE_sites(
  sites_to_run,
  run_pCap = FALSE,
  pCap_model_object_filepath = NULL,
  bugs_model_file,
  bugs_directory
)
}
\value{
A table with the format:
\itemize{
\item \strong{model_name}
\item \strong{site}
\item \strong{run_year}
\item \strong{week_fit}
\item \strong{location_fit}
\item \strong{parameter}
\item \strong{statistic}
\item \strong{value}
\item \strong{srjpedata_version}
}
}
\description{
Run all JPE sites.
}
\details{
This function automates running BT-SPAS-X for all JPE sites/run years.
}
