% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_multiple_bt_spas_x}
\alias{run_multiple_bt_spas_x}
\title{Call BT-SPAS-X model on all site/run years}
\usage{
run_multiple_bt_spas_x(
  bt_spas_x_bayes_params,
  bt_spas_x_input_data,
  lifestage,
  effort_adjust,
  mainstem_version,
  bugs_directory,
  debug_mode
)
}
\arguments{
\item{bt_spas_x_input_data}{data frame containing the same variables as
\code{?SRJPEdata::weekly_juvenile_abundance_model_data}}

\item{lifestage}{the lifestage for which you want tor run the model. One of \code{yearling}, \code{fry}, and \code{smolt}.}

\item{effort_adjust}{whether or not you want to use catch adjusted by effort}

\item{mainstem_version}{whether or not this is run on mainstem tributaries}

\item{bugs_directory}{where the \code{WinBUGS.exe} file can be found. Needs to end in \verb{/WinBUGS}}

\item{debug_mode}{whether you want to run \code{bugs} in debug mode.}

\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}. Can use \code{SRJPEmodel::bt_spas_x_bayes_params}.}
}
\value{
a list containing the results for every unique \code{site} and \verb{run year} combination
in the input data.
}
\description{
Call BT-SPAS-X model on all site/run years
}
\details{
This calls \code{run_single_bt_spas_x()} on all site and run year combinations
present in the input data.
}
