% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_to_spawner_model.R
\name{run_passage_to_spawner_model}
\alias{run_passage_to_spawner_model}
\title{Run Passage to Spawner (P2S) STAN Model}
\usage{
run_passage_to_spawner_model(
  stream_name,
  selected_covariate,
  extract_predicted_spawners = c(FALSE, TRUE)
)
}
\arguments{
\item{observed_adult_input:}{The exported data object \code{SRJPEdata::observed_adult_input}. See \code{?SRJPEdata::observed_adult_input} for
details on data structure.}

\item{adult_model_covariates_standard:}{The exported data object \code{SRJPEdata::adult_model_covariates_standard}. See \code{?SRJPEdata::adult_model_covariates_standard} for
details on data structure.}

\item{stream_name:}{The name of the stream for which you would like to run the P2S model. Can be
\verb{battle creek}, \verb{clear creek}, \verb{deer creek}, \verb{mill creek}, or (draft form) \verb{butte creek}.}

\item{selected_covariate:}{The environmental covariate you'd like to run the model for. Can be
either \code{wy_type} (water year type), \code{max_flow_std} (maximum flow), \code{gdd_std} (growing degree days),
\code{passage_index} (total upstream passage), \code{median_passage_timing_std} (median passage timing),
or \code{null_covar} (no environmental covariate).}

\item{extract_predicted_spawners:}{If \code{TRUE}, will only produce a tibble with median predicted spawners with upper (97.5\%)
and lower (2.5\%) confidence intervals. If \code{FALSE}, returns a list with all parameter estimates and the full \code{stanfit} object.
See \code{vignette("prep_environmental_covariates.Rmd", package = "SRJPEdata")} for more details.}
}
\value{
If \code{extracted_predicted_spawners == TRUE}, Returns a tibble containing the following variables:
\itemize{
\item \strong{stream} Stream name
\item \strong{year} Year observed data came from
\item \strong{median_predicted_spawners} Median predicted spawner value for a given stream and year
\item \strong{lcl} 2.5\% confidence interval for predicted spawner value
\item \strong{ucl} 97.5\% confidence interval for predicted spawner value
\item \strong{data_type} Data type: \code{P2S_predicted_spawners}
}

If \code{extracted_predicted_spawners == FALSE}, returns a list containing
\code{full_object}, the full fitted \code{stanfit} object (see \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{details}),
and \code{formatted_pars}, a formatted data table containing the following variables:
\itemize{
\item \strong{par_names} Parameter name
\item \strong{mean} Mean of the posterior distribution for a parameter
\item \strong{se_mean} Monte Carlo standard error for summary of all chains merged (see \href{https://mc-stan.org/rstan/reference/stanfit-method-summary.html}{details})
\item \strong{sd} Standard deviation of the posterior distribution for a parameter
\item \strong{\verb{2.5\%}} 2.5\% quantile of posterior distribution for a parameter.
\item \strong{\verb{25\%}} 25\% quantile of posterior distribution for a parameter.
\item \strong{\verb{50\%}} 50\% quantile of posterior distribution for a parameter.
\item \strong{\verb{75\%}} 75\% quantile of posterior distribution for a parameter.
\item \strong{\verb{97.5\%}} 97.5\% quantile of posterior distribution for a parameter.
\item \strong{n_eff} Effective sample size for a parameter
\item \strong{Rhat} Split Rhats for a parameter
\item \strong{stream} Stream name for the parameter estimate
\item \strong{year} Year observed data came from
}
}
\description{
This function takes in \code{SRJPEdata::observed_adult_input} and \code{SRJPEdata::adult_model_covariates_standard}
and runs the Passage to Spawner (P2S) model for a selected stream and environmental covariate. The function
will return a formatted list of parameter estimates and the full fitted \code{stanfit} object.
See \code{vignette("passage_to_spawner_submodel.Rmd", package = "SRJPEmodel")} for more details.
}
\seealso{
Other passage_to_spawner: 
\code{\link{compare_P2S_covariates_within_stream}()},
\code{\link{compare_P2S_model_covariates}()}
}
\concept{passage_to_spawner}
