% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_abundance.R
\name{run_single_bt_spas_x}
\alias{run_single_bt_spas_x}
\title{Call BT-SPAS-X on a single site/run year combination}
\usage{
run_single_bt_spas_x(
  bt_spas_x_bayes_params,
  bt_spas_x_input_data,
  site,
  run_year,
  life_stage,
  effort_adjust = c(T, F),
  mainstem_version = c(F, T),
  bugs_directory
)
}
\arguments{
\item{bt_spas_x_bayes_params:}{a list containing \code{number_mcmc}, \code{number_burnin}, \code{number_thin},
and \code{number_chains}}
}
\value{
TODO
}
\description{
Call BT-SPAS-X on a single site/run year combination
}
\details{
This function is called within \code{run_bt_spas_x()}
}
