% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{setup_azure_blob_backend}
\alias{setup_azure_blob_backend}
\title{Azure Blob Setup}
\usage{
setup_azure_blob_backend(
  storage_account,
  access_key = Sys.getenv("AZ_CONTAINER_ACCESS_KEY")
)
}
\arguments{
\item{storage_account}{string storage account to create connection to}

\item{access_key}{the access key that provides for the blob access. By default an env variable
with the name \code{AZ_CONTAINER_ACCESS_KEY} is retrived, optionally you can directly pass in a value.}
}
\value{
An \code{AzureStore} storage container object
}
\description{
Creates an \code{AzureStor} storage container object to be used by the \code{pins} package.
}
\keyword{internal}
