% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{store_model_fit}
\alias{store_model_fit}
\title{Store Model Fits}
\usage{
store_model_fit(
  con,
  storage_account,
  container_name,
  access_key,
  data,
  results_name,
  ...
)
}
\arguments{
\item{con}{a connection object to the database}

\item{storage_account}{an azure storage account}

\item{container_name}{container name in storage account (must exist)}

\item{access_key}{azure storage access key with write permissions}

\item{data}{the model result data object}

\item{...}{additional named arguments to be passed as metadata to the blob storage}

\item{result_name}{a name to identify the model results on azure storage}
}
\description{
Store a model fit object in a pre-configured Azure Blob Container and Database
}
