% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_model_code}
\alias{survival_model_code}
\title{Survival model code}
\format{
A named list with 2 elements:
\itemize{
\item \code{survival_CovWY}:
\item \code{survival_NoCov}:
}
}
\usage{
survival_model_code
}
\description{
A nested named list containing model code for versions of the survival models.
}
\keyword{datasets}
